Make Partners its own component and fix styling
