+++ dominion/dominion_ui/public/index.html
@@ -66,6 +66,7 @@
     <div id="dominionShowcase"></div>
     <div id="dominionData"></div>
     <div id="dominionAbout"></div>
+    <div id="dominionPartners"></div>
     <div id="dominionFooter"></div>
     <!--
       This HTML file is a template.
