+++ dominion/dominion_ui/src/components/Footer/index.js
@@ -5,7 +5,6 @@ import { Grid } from '@material-ui/core';
 import { withStyles } from '@material-ui/core/styles';
 
 import About from './About';
-import Partners from './Partners';
 
 const styles = () => ({
   root: {
@@ -21,9 +20,6 @@ function Footer({ classes }) {
       direction="row"
       alignItems="center"
     >
-      <Grid item xs={12}>
-        <Partners />
-      </Grid>
       <Grid item xs={12}>
         <About />
       </Grid>
