+++ dominion/templates/homepage.html
@@ -14,5 +14,5 @@ download data, and embed charts on your site.{% endblock %}
   <div id="dominionVideo"></div>
   <div id="dominionData"></div>
   <div id="dominionAbout"></div>
-
+  {% include 'blocks/partners.html' %}
 {% endblock content_container %}
